modularize query to a different module
