+++ src/pool.rs
@@ -4,6 +4,7 @@ use r2d2::Config;
 use r2d2_postgres::SslMode;
 use config::DbConfig;
 use database::{Database, DatabaseDDL, DatabaseDev};
+#[cfg(feature = "postgres")]
 use platform::Postgres;
 #[cfg(feature = "sqlite")]
 use platform::Sqlite;
@@ -22,6 +23,7 @@ use r2d2_sqlite::SqliteConnectionManager;
 
 /// the sql builder for each of the database platform
 pub enum Platform {
+    #[cfg(feature = "postgres")]
     Postgres(Postgres),
     #[cfg(feature = "sqlite")]
     Sqlite(Sqlite),
@@ -32,6 +34,7 @@ pub enum Platform {
 impl Platform {
     pub fn as_ref(&self) -> &Database {
         match *self {
+            #[cfg(feature = "postgres")]
             Platform::Postgres(ref pg) => pg,
             #[cfg(feature = "sqlite")]
             Platform::Sqlite(ref lite) => lite,
@@ -42,6 +45,7 @@ impl Platform {
 
     pub fn as_ddl(&self) -> &DatabaseDDL {
         match *self {
+            #[cfg(feature = "postgres")]
             Platform::Postgres(ref pg) => pg,
             #[cfg(feature = "sqlite")]
             Platform::Sqlite(ref lite) => lite,
@@ -52,6 +56,7 @@ impl Platform {
 
     pub fn as_dev(&self) -> &DatabaseDev {
         match *self {
+            #[cfg(feature = "postgres")]
             Platform::Postgres(ref pg) => pg,
             #[cfg(feature = "sqlite")]
             Platform::Sqlite(ref lite) => lite,
@@ -67,6 +72,7 @@ impl Deref for Platform {
     fn deref(&self) -> &Self::Target {
         debug!("using deref...");
         match *self {
+            #[cfg(feature = "postgres")]
             Platform::Postgres(ref pg) => pg,
             #[cfg(feature = "sqlite")]
             Platform::Sqlite(ref lite) => lite,
@@ -80,6 +86,7 @@ impl Deref for Platform {
 /// Postgres, Sqlite uses r2d2 connection manager,
 /// Mysql has its own connection pooling
 pub enum ManagedPool {
+    #[cfg(feature = "postgres")]
     Postgres(Pool<PostgresConnectionManager>),
     #[cfg(feature = "sqlite")]
     Sqlite(Pool<SqliteConnectionManager>),
@@ -95,6 +102,7 @@ impl ManagedPool {
             Some(config) => {
                 let platform: &str = &config.platform;
                 match platform {
+                    #[cfg(feature = "postgres")]
                     "postgres" => {
                         let manager = try!(PostgresConnectionManager::new(url, SslMode::None));
                         debug!("Creating a connection with a pool size of {}", pool_size);
@@ -138,6 +146,7 @@ impl ManagedPool {
     /// a conection is created here
     pub fn connect(&self) -> Result<Platform, DbError> {
         match *self {
+            #[cfg(feature = "postgres")]
             ManagedPool::Postgres(ref pool) => {
                 match pool.get() {
                     Ok(conn) => {
